Add hook generator password and more styles
