+++ src/components/johnsi15/PasswordGenerator.module.css
@@ -19,8 +19,12 @@
 }
 
 .password p {
-	opacity: .9;	
+	opacity: .6;	
 	font-size: 1.1rem;
+}
+
+.password .active {
+	opacity: .9;
 }
 
 .form {
